👷 setup CI/CD
